Fix creating a static web app (#650)
* Remove SubscriptionTreeItem, move create swa

* Remove AzureAccountTreeItemWithProjects
