+++ src/commands/createStaticWebApp/createStaticWebApp.ts
@@ -3,23 +3,39 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { AzExtFsExtra, IActionContext, ICreateChildImplContext } from '@microsoft/vscode-azext-utils';
-import { ProgressLocation, ProgressOptions, Uri, window } from 'vscode';
+import { WebSiteManagementClient } from '@azure/arm-appservice';
+import { LocationListStep, ResourceGroupCreateStep, ResourceGroupListStep, SubscriptionTreeItemBase, VerifyProvidersStep } from '@microsoft/vscode-azext-azureutils';
+import { AzExtFsExtra, AzureWizard, AzureWizardExecuteStep, AzureWizardPromptStep, IActionContext, ICreateChildImplContext, nonNullProp } from '@microsoft/vscode-azext-utils';
+import { ProgressLocation, ProgressOptions, Uri, window, workspace } from 'vscode';
 import { Utils } from 'vscode-uri';
 import { NodeConstants } from '../../detectors/node/nodeConstants';
 import { DetectorResults, NodeDetector } from '../../detectors/node/NodeDetector';
 import { VerifyingWorkspaceError } from '../../errors';
 import { ext } from '../../extensionVariables';
-import { SubscriptionTreeItem } from '../../tree/SubscriptionTreeItem';
+import { createWebSiteClient } from '../../utils/azureClients';
+import { getGitHubAccessToken } from '../../utils/gitHubUtils';
+import { gitPull } from '../../utils/gitUtils';
 import { localize } from '../../utils/localize';
 import { telemetryUtils } from '../../utils/telemetryUtils';
-import { getSubFolders, showNoWorkspacePrompt, tryGetWorkspaceFolder } from '../../utils/workspaceUtils';
+import { getSingleRootFsPath, getSubFolders, showNoWorkspacePrompt, tryGetWorkspaceFolder } from '../../utils/workspaceUtils';
+import { RemoteShortnameStep } from '../createRepo/RemoteShortnameStep';
+import { RepoCreateStep } from '../createRepo/RepoCreateStep';
+import { RepoNameStep } from '../createRepo/RepoNameStep';
+import { RepoPrivacyStep } from '../createRepo/RepoPrivacyStep';
+import { ApiLocationStep } from './ApiLocationStep';
+import { AppLocationStep } from './AppLocationStep';
+import { BuildPresetListStep } from './BuildPresetListStep';
+import { GitHubOrgListStep } from './GitHubOrgListStep';
 import { IStaticWebAppWizardContext } from './IStaticWebAppWizardContext';
+import { OutputLocationStep } from './OutputLocationStep';
 import { setWorkspaceContexts } from './setWorkspaceContexts';
+import { SkuListStep } from './SkuListStep';
+import { StaticWebAppCreateStep } from './StaticWebAppCreateStep';
+import { StaticWebAppNameStep } from './StaticWebAppNameStep';
 import { tryGetApiLocations } from './tryGetApiLocations';
 
 let isVerifyingWorkspace: boolean = false;
-export async function createStaticWebApp(context: IActionContext & Partial<ICreateChildImplContext> & Partial<IStaticWebAppWizardContext>, node?: SubscriptionTreeItem): Promise<void> {
+export async function createStaticWebApp(context: IActionContext & Partial<ICreateChildImplContext> & Partial<IStaticWebAppWizardContext>, node?: SubscriptionTreeItemBase): Promise<void> {
     if (isVerifyingWorkspace) {
         throw new VerifyingWorkspaceError(context);
     }
@@ -32,7 +48,7 @@ export async function createStaticWebApp(context: IActionContext & Partial<ICrea
     isVerifyingWorkspace = true;
     try {
         if (!node) {
-            node = await ext.rgApi.tree.showTreeItemPicker<SubscriptionTreeItem>(SubscriptionTreeItem.contextValue, context);
+            node = await ext.rgApi.tree.showTreeItemPicker<SubscriptionTreeItemBase>(SubscriptionTreeItemBase.contextValue, context);
         }
 
         await window.withProgress(progressOptions, async () => {
@@ -76,24 +92,80 @@ export async function createStaticWebApp(context: IActionContext & Partial<ICrea
         isVerifyingWorkspace = false;
     }
 
-    await node.createChildImpl2(context);
-    // // ext.rgApi.revealTreeItem()
-    // // void showSwaCreated(swaNode);
+    const client: WebSiteManagementClient = await createWebSiteClient([context, node.subscription]);
+    const wizardContext: IStaticWebAppWizardContext = { accessToken: await getGitHubAccessToken(), client, ...context, ...node.subscription };
 
-    // // only reveal SWA node when tree is visible to avoid changing their tree view just to reveal the node
-    // if (ext.rgApi.treeView.visible) {
-    //     const environmentNode: EnvironmentTreeItem | undefined = <EnvironmentTreeItem | undefined>(await swaNode.loadAllChildren(context)).find(ti => {
-    //         return ti instanceof EnvironmentTreeItem && ti.label === productionEnvironmentName;
-    //     });
-    //     environmentNode && await ext.treeView.reveal(environmentNode, { expand: true });
-    // }
+    const title: string = localize('createStaticApp', 'Create Static Web App');
+    const promptSteps: AzureWizardPromptStep<IStaticWebAppWizardContext>[] = [];
+    const executeSteps: AzureWizardExecuteStep<IStaticWebAppWizardContext>[] = [];
 
-    // void postCreateStaticWebApp(swaNode);
-    // return swaNode;
+    if (!context.advancedCreation) {
+        wizardContext.sku = SkuListStep.getSkus()[0];
+        executeSteps.push(new ResourceGroupCreateStep());
+    } else {
+        promptSteps.push(new ResourceGroupListStep());
+    }
 
+    promptSteps.push(new StaticWebAppNameStep(), new SkuListStep());
+    const hasRemote: boolean = !!wizardContext.repoHtmlUrl;
+
+    // if the local project doesn't have a GitHub remote, we will create it for them
+    if (!hasRemote) {
+        promptSteps.push(new GitHubOrgListStep(), new RepoNameStep(), new RepoPrivacyStep(), new RemoteShortnameStep());
+        executeSteps.push(new RepoCreateStep());
+    }
+
+    // hard-coding locations available during preview
+    // https://github.com/microsoft/vscode-azurestaticwebapps/issues/18
+    const locations = [
+        'Central US',
+        'East US 2',
+        'East Asia',
+        'West Europe',
+        'West US 2'
+    ];
+
+    const webProvider: string = 'Microsoft.Web';
+
+    LocationListStep.setLocationSubset(wizardContext, Promise.resolve(locations), webProvider);
+    LocationListStep.addStep(wizardContext, promptSteps, {
+        placeHolder: localize('selectLocation', 'Select a region for Azure Functions API and staging environments'),
+        noPicksMessage: localize('noRegions', 'No available regions.')
+    });
+
+    promptSteps.push(new BuildPresetListStep(), new AppLocationStep(), new ApiLocationStep(), new OutputLocationStep());
+
+    executeSteps.push(new VerifyProvidersStep([webProvider]));
+    executeSteps.push(new StaticWebAppCreateStep());
+
+    const wizard: AzureWizard<IStaticWebAppWizardContext> = new AzureWizard(wizardContext, {
+        title,
+        promptSteps,
+        executeSteps,
+        showLoadingPrompt: true
+    });
+
+    wizardContext.telemetry.properties.gotRemote = String(hasRemote);
+    wizardContext.fsPath = wizardContext.fsPath || getSingleRootFsPath();
+    wizardContext.telemetry.properties.numberOfWorkspaces = !workspace.workspaceFolders ? String(0) : String(workspace.workspaceFolders.length);
+
+    await wizard.prompt();
+
+    // const newStaticWebAppName: string = nonNullProp(wizardContext, 'newStaticWebAppName');
+
+    if (!context.advancedCreation) {
+        wizardContext.newResourceGroupName = await wizardContext.relatedNameTask;
+    }
+
+    await wizard.execute();
+
+    // const swa: WebSiteManagementModels.StaticSiteARMResource = nonNullProp(context, 'staticWebApp');
+    await gitPull(nonNullProp(wizardContext, 'repo'));
+
+    await ext.rgApi.tree.refresh(context);
     void node.refresh(context);
 }
 
-export async function createStaticWebAppAdvanced(context: IActionContext, node?: SubscriptionTreeItem): Promise<void> {
+export async function createStaticWebAppAdvanced(context: IActionContext, node?: SubscriptionTreeItemBase): Promise<void> {
     return await createStaticWebApp({ ...context, advancedCreation: true }, node);
 }
